+++ packages/perseus/src/types.ts
@@ -506,9 +506,9 @@ export type PerseusDependencies = {
  *
  * Prefer using this type over `PerseusDependencies` when possible.
  */
-export type PerseusDependenciesV2 = {
+export interface PerseusDependenciesV2 {
     analytics: {onAnalyticsEvent: AnalyticsEventHandlerFn};
-};
+}
 
 /**
  * APIOptionsWithDefaults represents the type that is provided to all widgets.